Complete completeness proof
This completes the completeness proof!  The key outstanding problem was
the Merkle proofs were not organised correctly.  They needed to be in a
reverse topological order and, furthermore, required path signs.

With all this in place, it was not so hard to get the completeness proof
through.  In fact, Dafny resolved it automatically using induction.