Working on Lagrange Polynomials
The basic mechanism for encoding into polynomials works, but I've hit a
roadblock with the Lagrange Interpolation.  Specifically, I cannot
construct an actual polynomial from a blob of data.  The problem relates
to division, and that I cannot divide a polynomial accurately.

There are a few options.  One option would be to use a polynomial
representation with rational coefficients.  That would work because the
denominator of the lagrange polynomial is always a constant.

The other (and more obvious option) would be to forget trying to encode
polynomials over integers and, instead, encode them over fields.  In
particular, prime fields where every element has an inverse would of
course allow me to compute an actual polynomial.  This of course would
better reflect what we are actually trying to do here.